+++ core/PSRemoteOperations.psm1
@@ -3,10 +3,6 @@
 . $PSScriptRoot\..\functions.ps1
 . $PSScriptRoot\..\private.ps1
 
-#if running Windows on Powershell Core, include another command
-if ($isWindows) {
-    . $PSScriptRoot\..\windows\Register-PSRemoteOperationWatcher.ps1
-}
 #endregion
 
 #add default properties for the custom result object